+++ QMP_G7/src/main/java/dominio/prendas/ReqClimaticos.java
@@ -4,8 +4,8 @@ import servicios.Clima;
 
 public class ReqClimaticos {
 
-  private Integer temperaturaMaxF;
-  private Integer temperaturaMinF;
+  private Float temperaturaMaxF;
+  private Float temperaturaMinF;
 
   // Por el momento solo valida la temperatura, que es lo que pide el enunciado.
   // Realizar otras validaciones aqui si se quieren agregar.
@@ -15,11 +15,11 @@ public class ReqClimaticos {
         unClima.getTemperatureF() > this.temperaturaMinF;
   }
 
-  public void setTemperaturaMinF(Integer temperaturaMinF) {
+  public void setTemperaturaMinF(Float temperaturaMinF) {
     this.temperaturaMinF = temperaturaMinF;
   }
 
-  public void setTemperaturaMaxF(Integer temperaturaMaxF) {
+  public void setTemperaturaMaxF(Float temperaturaMaxF) {
     this.temperaturaMaxF = temperaturaMaxF;
   }
